Automatically backed up by Learn

+++ db/migrate/20190829185559_users.rb
@@ -3,6 +3,6 @@ class Users < ActiveRecord::Migration[5.1]
     create_table :users do |t|
       t.string :username
       t.string :password
-      t.integer :balance
+      t.decimal :balance
   end
 end
